make JSONP conditional

+++ install.sh
@@ -12,12 +12,19 @@ unzip -q ./geoserver-tmp/geoserver.zip -d ./webapps/geoserver
 rm -rf geoserver-tmp
 rm geoserver.zip
 
+echo "Enable JSONP? [y/n] (n): ";
+read answer
+
 # enbale JSONP
+if [ "$answer" == "y" ] then
+
 echo "Enabling JSONP output...";
 # sed is insane
 sed -e ':a' -e 'N' -e '$!ba' -e 's/<!--\n*[ ]*\(<[^>]*>\n*[ ]*<[^>]*>ENABLE_JSONP<\/[^>]*>[ ]*\n*[ ]*<[^>]*>[^<]*<\/[^>]*>\n*[ ]*<[^>]*>[ ]*\n*[ ]*\)-->/\1/g' ./webapps/geoserver/WEB-INF/web.xml > ./webapps/geoserver/WEB-INF/w.xml
 cat ./webapps/geoserver/WEB-INF/w.xml > ./webapps/geoserver/WEB-INF/web.xml
 rm ./webapps/geoserver/WEB-INF/w.xml
+
+fi
 
 echo "Remove pom.xml to avoid Maven builds? [y/n] (n): ";
 read answer
